refact marathon v1v2
